Cloud/infrastructure/terraform/core/cloud_run/requirements.txt
Flask==2.3.3
google-cloud-compute==1.23.0
gunicorn==21.2.0
requests==2.31.0 